add feed and community posts page components
